delete user doesnt take param id

+++ src/Endpoints/DeleteUser.php
@@ -4,20 +4,21 @@ require_once __DIR__ . '/../Service/UserService.php';
 
 $userService = new UserService();
 
-// Assuming you have a way to get $id, $email, and $username from the request
-$id = isset($id) ? intval($id) : 0;
-$email = isset($data['email']) ? $data['email'] : '';
-$username = isset($data['username']) ? $data['username'] : '';
+$inputData = file_get_contents("php://input");
+$data = json_decode($inputData, true);
 
-$userDto = new UserDto($id, $email, $username);
+if (isset($data['id'])) {
+    $userDto = new UserDto($data['id'], '', ''); // Create UserDto with just the ID
+    $success = $userService->deleteUser($userDto);
 
-// Delete the user
-$success = $userService->deleteUser($userDto);
-
-if ($success) {
-    header("Content-Type: application/json");
-    echo json_encode(array("message" => "User deleted successfully"));
+    if ($success) {
+        header("Content-Type: application/json");
+        echo json_encode(array("message" => "User deleted successfully"));
+    } else {
+        http_response_code(404);
+        echo json_encode(array("message" => "User not found or couldn't be deleted"));
+    }
 } else {
-    http_response_code(404);
-    echo json_encode(array("message" => "User not found or couldn't be deleted"));
+    http_response_code(400);
+    echo json_encode(array("message" => "Missing user ID in request body"));
 }
